- modified chess field position's properties similar to chess piece's properties (binary data accessed with getters / setters) - tested the chess field position getters / setters

+++ Chess.Lib/ChessPiece.cs
@@ -109,7 +109,7 @@ namespace Chess.Lib
         /// </summary>
         public ChessFieldPosition Position
         {
-            get { return new ChessFieldPosition(_hashCode & BITS_OF_POSITION); }
+            get { return new ChessFieldPosition((byte)(_hashCode & BITS_OF_POSITION)); }
             set { _hashCode = (short)((_hashCode & ~BITS_OF_POSITION) | value.GetHashCode()); }
         }
